+++ services/engine/codec/video/hcodec/hdecoder.cpp
@@ -162,7 +162,7 @@ int32_t HDecoder::UpdateOutPortFormat()
     }
     outputFormat_->PutIntValue(MediaDescriptionKey::MD_KEY_WIDTH, w);
     outputFormat_->PutIntValue(MediaDescriptionKey::MD_KEY_HEIGHT, h);
-    outputFormat_->PutIntValue(MediaDescriptionKey::MD_KEY_PIXEL_FORMAT, configuredFmt_.innerFmt);
+    outputFormat_->PutIntValue(MediaDescriptionKey::MD_KEY_PIXEL_FORMAT, static_cast<int32_t>(configuredFmt_.innerFmt));
 
     sharedBufferFormat_ = { w, h, def.format.video.nStride, OMX_VIDEO_CodingUnused, configuredFmt_ };
     return AVCS_ERR_OK;
